Changed to float and added test

+++ NetworkRouting/ArrayImpl.cs
@@ -9,14 +9,14 @@ namespace NetworkRouting
     class ArrayImpl : IDijkstraShortestPathQueue
     {
         // The queue
-        int[] queue = null;
+        float[] queue = null;
 
         // Helper members
         int size = 0;
 
         // Other
         private int QUEUE_DEQUEUED_PLACEHOLDER = -1;
-        private int QUEUE_NOT_INITIALIZED_PLACEHOLDER = Int32.MaxValue;
+        private float QUEUE_NOT_INITIALIZED_PLACEHOLDER = float.MaxValue;
 
         public ArrayImpl(int numPoints)
         {
@@ -28,7 +28,7 @@ namespace NetworkRouting
             return this.size;
         }
 
-        void IDijkstraShortestPathQueue.insert(int index, int weight)
+        void IDijkstraShortestPathQueue.insert(int index, float weight)
         {
             queue[index] = weight;
             this.size++;
@@ -37,7 +37,7 @@ namespace NetworkRouting
         int IDijkstraShortestPathQueue.deleteMin()
         {
             int minIndex = 0;
-            int minValue = queue[minIndex];
+            float minValue = queue[minIndex];
             for(int i = 0; i < queue.Length; i++)
             {
                 if(queue[i] < minValue)
@@ -52,7 +52,7 @@ namespace NetworkRouting
             return minIndex;
         }
 
-        void IDijkstraShortestPathQueue.decreaseKey(int index, int newWeight)
+        void IDijkstraShortestPathQueue.decreaseKey(int index, float newWeight)
         {
             // If the key is being decreased for the first time,
             // it is the same as inserting a new element into the queue.
